# Improve performance of `auto_grader_spec` test